==> Emotional Meta Memory - Analysis Only/4_Trial_2_Count_Function.Rmd <==
---
title: "Trial 2 Count"
output: html_document
---

#Making R-Version of the 'Trial 2 Count' Function used by the Meta Cognitive Model implemented in MatLab (https://github.com/metacoglab/HMeta-d/blob/master/Matlab/trials2counts.m)
```{r setup}
#To install rJags (https://gist.github.com/casallas/8411082)
## Load Libraries
library(tidyverse)
library(magrittr)
library(reshape2)
library(rjags)
library(coda)
library(lattice)
library(broom)
library(ggpubr)
library(ggmcmc)
```



#The Function
##It converts trial by trial experimental information for N trials into response counts.

trials2counts <- function(stimID, response, rating,nRatings, padAmount = 0,padCells=0)
    
INPUTS
stimID - Input for expected Keypress in List Format
1xN vector. 
stimID(i) = 0 --> stimulus on i'th trial was S1.
stimID(i) = 1 --> stimulus on i'th trial was S2.
                       
response - Input for Actual Keypress in List Format
1xN vector. 
response(i) = 0 --> response on i'th trial was 'S1'.
response(i) = 1 --> response on i'th trial was 'S2'.

rating - Input for Confidence Rating in List Format
1xN vector. 
rating(i) = X --> rating on i'th trial was X.
X must be in the range 1 <= X <= nRatings.

nRatings - Total # of available subjective ratings available for the subject. 
e.g. if subject can rate confidence on a scale of 1-4, then nRatings = 4

cellpad: 
if set to 1, each response count in the output has the value of
1/(2*nRatings) added to it. This is desirable if trial counts of
0 interfere with model fitting.
if set to 0, trial counts are not manipulated and 0s may be present.
Default value is 0.

OUTPUTS
nR_S1, nR_S2
these are vectors containing the total number of responses in
each response category, conditional on presentation of S1 and S2.

e.g. if nR_S1 = [100 50 20 10 5 1], 
then when stimulus S1 was
presented, the subject had the following response counts:
responded S1, rating=3 : 100 times
responded S1, rating=2 : 50 times
responded S1, rating=1 : 20 times
responded S2, rating=1 : 10 times
responded S2, rating=2 : 5 times
responded S2, rating=3 : 1 time

```{r}
trials2counts <- function(stimID, response, rating,nRatings, padAmount = 0,padCells=0){

  nR_S1 <- list()
  nR_S2 <- list()

    if (padAmount == 0){
        padAmount = 1/(2*nRatings)}
  
  # S1 responses
    for (r in nRatings:1){
        cs1 <- 0
        cs2 <- 0
        for (i in 1:length(stimID)){
            s = stimID[i]
            x = response[i]
            y = rating[i]
            
            if ((s==0) & (x==0) & (y==r)){
                (cs1 <- cs1+1)}
            if ((s==1) & (x==0) & (y==r)){
                (cs2 <- cs2+1)}
            }
        nR_S1 <- append(nR_S1,cs1)
        nR_S2 <- append(nR_S2,cs2)
        }
    
    # S2 responses
    for (r in 1:nRatings){
     cs1 <- 0
        cs2 <- 0
        for (i in 1:length(stimID)){
            s = stimID[i]
            x = response[i]
            y = rating[i]
            
            if ((s==0) & (x==1) & (y==r)){
                (cs1 <- cs1+1)}
            if ((s==1) & (x==1) & (y==r)){
                (cs2 <- cs2+1)}
            }
        nR_S1 <- append(nR_S1,cs1)
        nR_S2 <- append(nR_S2,cs2)
        }
  
    
    # pad response counts to avoid zeros
      nR_S1 <- as.numeric(nR_S1)
      nR_S2 <- as.numeric(nR_S2)
  if (padCells == 1){
  nR_S1 <- lapply(nR_S1,FUN= function(x) x+padAmount)
  nR_S2 <- lapply(nR_S2,FUN= function(x) x+padAmount)}
  
    # Combine into lists
  newlist <- list(nR_S1,nR_S2)
    }
```
